[MID-88] move query extensions to separate file
